% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{addinLoadTest}
\alias{addinLoadTest}
\title{Addin To Setup Load Test}
\usage{
addinLoadTest()
}
\description{
Addin To Setup Load Test
}
\details{
Runs a Shiny Gadget that helps users build a load test report -
  similar to knitting the parameterized report, but with more helpful
  defaults
}
