% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R, R/plotting.R
\name{plot_concurrency_time_by_eventtype}
\alias{plot_concurrency_time_by_eventtype}
\alias{plot_gantt}
\alias{plot_gantt_session}
\alias{plot_gantt_duration}
\alias{plot_gantt_latency}
\alias{plot_http_latency}
\alias{plot_websocket_latency}
\alias{plot_timeline}
\title{Analysis plots}
\usage{
plot_concurrency_time_by_eventtype(df)

plot_gantt(df)

plot_gantt_session(df)

plot_gantt_duration(df, cutoff = 10)

plot_gantt_latency(df)

plot_http_latency(df, cutoff = 10)

plot_websocket_latency(df, cutoff = 10)

plot_timeline(df)
}
\arguments{
\item{df}{dataframe returned from \code{\link{tidy_loadtest}}}
}
\description{
Analysis plots
}
