% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{createLoadTestReport}
\alias{createLoadTestReport}
\title{Create load test report from template}
\usage{
createLoadTestReport(dir = getwd(), name = "load_test_template.Rmd")
}
\arguments{
\item{dir}{Destination directory for the load test template}

\item{name}{New name for the load test template. Must include the .Rmd extension.}
}
\description{
Create load test report from template
}
\details{
Copies the load test template from \code{shinyloadtest} to \code{dir/name}
}
