% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-log.R
\name{createLog}
\alias{createLog}
\title{Create log file from directory containing run logs}
\usage{
createLog(logsDir)
}
\arguments{
\item{logsDir}{- directory containing run logs, see \code{loadtest}}
}
\value{
data frame with raw event logs from successful tests
}
\description{
Create log file from directory containing run logs
}
