% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-log.R
\name{getConcurrentOverTest}
\alias{getConcurrentOverTest}
\title{Get concurrent connections over the test duration}
\usage{
getConcurrentOverTest(eventLog)
}
\arguments{
\item{eventLog}{Data frame of events returned from \code{\link{createLog}}}
}
\value{
A data frame with time stamps and the number of concurrent
  connections for each time stamp
}
\description{
Get concurrent connections over the test duration
}
