% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-log.R
\name{getEventInterval}
\alias{getEventInterval}
\title{Get duration for simple event pairs}
\usage{
getEventInterval(eventLog, startEvent, endEvent)
}
\arguments{
\item{eventLog}{Data frame of events returned from \code{\link{createLog}}}

\item{startEvent}{The first event in the pair, i.e. "REQ_HOME"}

\item{endEvent}{The second event in the pair, i.e. "WS_CLOSE"}
}
\value{
A data frame containing the duration of the event pair per connection
}
\description{
Get duration for simple event pairs
}
