% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-log.R
\name{getPageLoadTimes}
\alias{getPageLoadTimes}
\title{Get page load times}
\usage{
getPageLoadTimes(eventLog)
}
\arguments{
\item{eventLog}{Data frame of events returned from \code{\link{createLog}}}
}
\value{
Data frame containing the page load time per connection
}
\description{
Get page load times
}
