% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-log.R
\name{getSetInputTimes}
\alias{getSetInputTimes}
\title{Get event duration for user interactions with inputs}
\usage{
getSetInputTimes(eventLog)
}
\arguments{
\item{eventLog}{eventLog Data frame of events returned from either
\code{\link{loadTest}} or \code{\link{getSuccesses}}}
}
\value{
Data frame containing a unique input id and event duration
}
\description{
Get event duration for user interactions with inputs
}
