% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-log.R
\name{getSuccesses}
\alias{getSuccesses}
\title{Get successes from event log}
\usage{
getSuccesses(eventLog)
}
\arguments{
\item{eventLog}{Returned from \code{\link{loadTest}}. The \code{eventLog} is
a list composed of data frames (from successful tests) and potnetially errors.}
}
\description{
Get successes from event log
}
\details{
Given an \code{eventLog} containing successesful test results (data
  frames) and error messages, this function returns the successful results in
  a collapsed data frame. See \code{\link{getErrors}}
}
