% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-log.R
\name{getTestDurations}
\alias{getTestDurations}
\title{Get test duration}
\usage{
getTestDurations(eventLog)
}
\arguments{
\item{eventLog}{Data frame of events returned from \code{\link{createLog}}}
}
\value{
Data frame containing the test duration per connection.
}
\description{
Get test duration
}
