% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{guessTestDuration}
\alias{guessTestDuration}
\title{Guess Test Duration}
\usage{
guessTestDuration(test_file)
}
\arguments{
\item{test_file}{full path and filename for .R file containing the load test
script, generated from \code{shinytest::recordTest(load_mode = TRUE)}.}
}
\value{
Expected run time for a single connection in seconds.
}
\description{
Guess Test Duration
}
\details{
Parses the \code{test_file} to determine the run time for a single connection.
}
