% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{hist_loadtimes}
\alias{hist_loadtimes}
\title{Histogram of Page Load Times}
\usage{
hist_loadtimes(df, max_load_time = 5)
}
\arguments{
\item{df}{A tidy data frame of test results. See \code{\link{tidy_loadtest}}}

\item{max_load_time}{The amount of time users will wait for the page to load
when first requesting the app.}
}
\description{
Histogram of Page Load Times
}
