% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{load_runs}
\alias{load_runs}
\title{Create Tidy Load Test Results}
\usage{
load_runs(..., verbose = TRUE)
}
\arguments{
\item{...}{Key-value pairs where the key is the desired name for the test and the
value is a path to the test result directory.}

\item{verbose}{Whether or not to print progress for reading loadtest directories}
}
\value{
A tidy data frame with the test result data. Each row is an event. Columns include
identifiers and timing information for the event.
}
\description{
The \code{shinycannon} tool creates a directory of log files for
each load test. This function translates one or more test result
directories into a tidy data frame.
}
\examples{
\dontrun{
  load_runs(
     `1 core` = 'results/run-1/',
     `2 cores` = 'results/run-2/'
  )
}
}
