% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_output.R
\name{make_report}
\alias{make_report}
\title{Make shinyloadtest Report}
\usage{
make_report(df)
}
\arguments{
\item{df}{data.frame returned from \code{\link{tidy_loadtest}}}
}
\description{
Make shinyloadtest Report
}
