% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_concurrency_time}
\alias{plot_concurrency_time}
\title{Concurrency Over Time}
\usage{
plot_concurrency_time(df, labels = NULL)
}
\arguments{
\item{df}{A tidy data frame of test results. See \code{\link{tidy_loadtest}}}

\item{labels}{A vector of labels to include}
}
\description{
Concurrency Over Time
}
