% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_loadtest}
\alias{plot_loadtest}
\alias{plot_time_boxplot}
\alias{plot_concurrency_time}
\alias{plot_timeline}
\alias{plot_hist_loadtimes}
\alias{plot_gantt}
\alias{plot_gantt_session}
\alias{plot_gantt_duration}
\alias{plot_gantt_latency}
\alias{plot_http_latency}
\alias{plot_websocket_latency}
\title{Plotting outputs for tidy_loadtest}
\usage{
plot_time_boxplot(df, labels = NULL)

plot_concurrency_time(df, labels = NULL)

plot_timeline(df)

plot_hist_loadtimes(df, max_load_time = 5)

plot_gantt(df)

plot_gantt_session(df)

plot_gantt_duration(df, cutoff = 10)

plot_gantt_latency(df)

plot_http_latency(df, cutoff = 10)

plot_websocket_latency(df, cutoff = 10)
}
\arguments{
\item{df}{data frame returned from \code{\link{tidy_loadtest}}}

\item{labels}{A vector of labels to include.  If none are supplied, all labels will be used.}

\item{max_load_time}{The amount of time users will wait for the page to load
when first requesting the app.}

\item{cutoff}{Where to draw a horizontal or vertical line to display a resonable cutoff line for requests.}
}
\description{
Many different plotting routines to display different loadtest information.
}
\section{Functions}{
\itemize{
\item \code{plot_time_boxplot}: Box plot of load times for each event in each run

\item \code{plot_concurrency_time}: Time on concurrency for each event for each run

\item \code{plot_timeline}: Event waterfall for each session within each run

\item \code{plot_hist_loadtimes}: Histogram of page load times

\item \code{plot_gantt}: Gantt chart of event duration for each session within each run

\item \code{plot_gantt_session}: Event gantt chart of each user session within each run

\item \code{plot_gantt_duration}: Event gantt chart of fastest to slowest session times within each run

\item \code{plot_gantt_latency}: Stacked bar chart of event duration for each session within each run

\item \code{plot_http_latency}: Bar chart of total HTTP latency for each session within each run

\item \code{plot_websocket_latency}: Bar chart of maximum calculation (websocket) latency for each session within each run
}}

