% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll.R
\name{plot_metrics}
\alias{plot_metrics}
\title{Plot server-reported metrics}
\usage{
plot_metrics(metrics)
}
\arguments{
\item{metrics}{server reported metrics from \code{\link{poll}}}
}
\description{
Plot server-reported metrics
}
