% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_time_boxplot}
\alias{plot_time_boxplot}
\title{Boxplot of Event Duration}
\usage{
plot_time_boxplot(df, labels = NULL)
}
\arguments{
\item{df}{A tidy data frame of test results. See \code{\link{tidy_loadtest}}}

\item{labels}{A vector of labels to include}
}
\description{
Boxplot of Event Duration
}
