% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll.R
\name{poll}
\alias{poll}
\title{Poll server(s) during load test}
\usage{
poll(servers, appName, duration_sec, platform)
}
\arguments{
\item{servers}{vector of servers supporting app i.e.
c('http://connect1.example.com', 'http://connect2.example.com')}

\item{appName}{name of the deployed application}

\item{duration_sec}{duration of polling in seconds}

\item{platform}{one of: 'connect'}
}
\value{
generates a real time plot of concurrent connections, aggregated
  load, and aggregated RAM used by the app
}
\description{
Poll server(s) during load test
}
