% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll.R
\name{poll_multiple}
\alias{poll_multiple}
\title{Responsible for polling the desired metrics across all servers and
aggregating the results}
\usage{
poll_multiple(endpoints, metric)
}
\arguments{
\item{endpoints}{see \code{\link{resolve_endpoints}}}

\item{metric}{one of: ram, cpu, concurrent}
}
\value{
data frame
}
\description{
Responsible for polling the desired metrics across all servers and
aggregating the results
}
