% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll.R
\name{poll_once}
\alias{poll_once}
\title{Switches between metrics for one server}
\usage{
poll_once(endpoint, metric)
}
\arguments{
\item{endpoint}{a single endpoint, see \code{\link{resolve_endpoints}}}

\item{metric}{one of ram, cpu, concurrent}
}
\value{
A dataframe row with timestamp, metric, and metric value for one server
}
\description{
Switches between metrics for one server
}
