% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-recorder.R
\name{record_session}
\alias{record_session}
\title{Record a Session for Load Test}
\usage{
record_session(target_app_url, host = "127.0.0.1", port = 8600,
  output_file = "recording.log", open_browser = TRUE)
}
\arguments{
\item{target_app_url}{The URL of the deployed application}

\item{host}{The host where the proxy will run. Usually localhost is used.}

\item{port}{The port for the proxy. Default is 8600. Change this default if
port 8600 is used by another service.}

\item{output_file}{The name for the generated log file.}

\item{open_browser}{Whether to open a browser on the proxy
(default=\code{TRUE}) or not (\code{FALSE}).}
}
\value{
Creates a recording file that can be used to drive a load test.
}
\description{
Record a Session for Load Test
}
\details{
This function creates a reverse proxy listening at \code{host:port}
  that stands in front of \code{target_app_url}. As the user interacts with
  the app, the actions and responses are recorded in a file that can be
  replayed later. By default, a web browser is opened automatically so that
  you can begin recording immediately, without manually starting a browser
  and navigating to \code{host:port}.
}
