% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyRecorder.R
\name{recordSession}
\alias{recordSession}
\title{Title}
\usage{
recordSession(targetAppUrl, host = "0.0.0.0", port = 8600,
  outputFile = "recording.log")
}
\arguments{
\item{outputFile}{}
}
\description{
Title
}
