% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll.R
\name{resolve_endpoints}
\alias{resolve_endpoints}
\title{Create endpoints object from inputs}
\usage{
resolve_endpoints(servers, appName, platform)
}
\arguments{
\item{servers}{server base url, see \code{\link{poll}}}

\item{appName}{app name, see \code{\link{poll}}}

\item{platform}{one of: 'connect'}
}
\value{
a list where each server has an endpoint, which is a list with server: url,
  appName: appName. The endpoint's class is the platform, used to
  dispatch the appropriate polling methods
}
\description{
Create endpoints object from inputs
}
