% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-load-driver.R
\docType{data}
\name{ShinyLoadDriver}
\alias{ShinyLoadDriver}
\title{Class that extends ShinyDriver from \code{shinytes}t package for load testing.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ShinyLoadDriver
}
\description{
Class that extends ShinyDriver from \code{shinytes}t package for load testing.
}
\section{Usage}{

app <- ShinyLoadDriver$new(url, connectionId)
app$snapshotInit("myloadtest")
app$setInputs(...)
app$snapshot()
app$stop()
app$getEventLog()
}

\section{Arguments}{

\describe{
  \item{path}{The url of a deployed Shiny app starting
   with 'http://' or https://'}
  \item{connectionId}{Identifies the current connection,
  typically a sequential id}
  \item{...}{\code{ShinyLoadDriver} inherits the other
  arguments from \code{ShinyDriver}}
 }
}

\section{Details}{


\code{app$new()} creates a \code{ShinyLoadDriver} object. It opens
a session in phantomJS connected to the Shiny app deployed at \code{path}.
Because ShinyLoadDriver connects to a deployed application, not all of the
methods from \code{ShinyDriver} are available.

\code{app$setInputs()} sets an input value in the Shiny app and
records the amount of time before either an output is updated or a timeout
occurs. The timeout value, in milliseconds, can be specified using
\code{setInputs(..., timeout_ = 3000)}. \code{setInputs()} accepts at most 1
input at a time.

\code{app$snapshot()} takes a screenshot of the application and saves it to
a sub-directory of the current working directory. The subdirectory is named
based on \code{app$snapshotInit("directoryName")}. Screenshots are named
\code{connectionId_snapshotCount.png} where \code{snapshotCount} is a counter
incrementing with each screenshot taken by the \code{ShinyLoadDriver}.

\code{app$getEventLog()} returns an event log with information on the test.
}

\keyword{datasets}
