% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyloadtest-package.R
\docType{package}
\name{shinyloadtest-package}
\alias{shinyloadtest}
\alias{shinyloadtest-package}
\title{shinyloadtest: Load Test Shiny Applications}
\description{
Assesses the number of concurrent users 'shiny' applications are capable of supporting, and for directing application changes in order to support a higher number of users. Provides facilities for recording 'shiny' application sessions, playing recorded sessions against a target server at load, and analyzing the resulting metrics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/shinyloadtest/}
  \item \url{https://github.com/rstudio/shinyloadtest}
  \item Report bugs at \url{https://github.com/rstudio/shinyloadtest/issues}
}

}
\author{
\strong{Maintainer}: Barret Schloerke \email{barret@posit.co} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})

Authors:
\itemize{
  \item Alan Dipert
  \item Barbara Borges
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
