% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slt_demo_data_16}
\alias{slt_demo_data_16}
\title{Example metrics for a 16-user load test}
\format{
A data frame with 5402 rows and 12 variables:
\describe{
\item{run}{Name of the run}
\item{session_id}{simulated session identifier, 0-based}
\item{user_id}{simulated user identifier, 0-based}
\item{iteration}{user session identifier, 0-based}
\item{input_line_number}{recording line number associated with event}
\item{event}{type of the event}
\item{start}{time the event started, in seconds, relative to the time at which all simulated users were running.}
\item{end}{time the event ended, in seconds, relative to the time at which all simulated users were running}
\item{time}{event duration, in seconds}
\item{concurrency}{number of events that happened at the same time as this one}
\item{maintenance}{whether this event occurred before or after all simulated users were running}
\item{label}{event-specific text label}
}
}
\usage{
slt_demo_data_16
}
\description{
An example dataset like that returned by \code{\link{load_runs}}, but without the \code{json} variable for portability. Contains latency data for 5402 \code{shinycannon} events suitable for passing to \code{\link{shinyloadtest_report}}.
}
\keyword{datasets}
