% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{tidy_loadtest}
\alias{tidy_loadtest}
\title{Create Tidy Load Test Results}
\usage{
tidy_loadtest(
     test1 = 'results/test-1/',
     test2 = 'results/test-2/'
  )
}
\arguments{
\item{...}{Key-value pairs where the key is the desired name for the test and the
value is a path to the test result directory.}
}
\value{
A tidy data frame with the test result data. Each row is an event. Columns include
   identifiers and timing information for the event.
}
\description{
The \code{shinycannon} tool creates a directory of log files for
  each load test. This function translates one or more test result
  directories into a tidy data frame.
}
