% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tidy-eval.R
\name{tidyeval}
\alias{tidyeval}
\alias{quo}
\alias{quos}
\alias{enquo}
\alias{sym}
\alias{syms}
\alias{ensym}
\alias{expr}
\alias{exprs}
\alias{enexpr}
\alias{quo_name}
\title{Tidy eval helpers}
\description{
These functions provide tidy eval-compatible ways to capture
symbols (\code{sym()}, \code{syms()}, \code{ensym()}), expressions (\code{expr()},
\code{exprs()}, \code{enexpr()}), and quosures (\code{quo()}, \code{quos()}, \code{enquo()}).
To learn more about tidy eval and how to use these tools, read
\url{http://rlang.tidyverse.org/articles/tidy-evaluation.html}
}
\keyword{internal}
