% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-log.R
\name{trimRampUp}
\alias{trimRampUp}
\title{Remove connections during ramp up from log}
\usage{
trimRampUp(eventLog, rampUpConns)
}
\arguments{
\item{eventLog}{A data frame with event information as returned from
\code{getSuccesses}}

\item{rampUpConns}{Number of connections to remove.}
}
\value{
Trimmed \code{eventLog}
}
\description{
Remove connections during ramp up from log
}
\details{
\code{rampUpConns} connections are removed from the beginning of the
  event log. Trimming these connections can help stabalize the results of the
  load test.
}
