% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R, R/plotting.R
\name{plot_concurrency_time_by_eventtype}
\alias{plot_concurrency_time_by_eventtype}
\alias{plot_gantt}
\alias{plot_gantt_session}
\alias{plot_gantt_iteration}
\alias{plot_gantt_duration}
\alias{plot_gantt_latency}
\alias{plot_timestamp_time}
\alias{plot_timeline}
\alias{plot_timeline_stacked}
\alias{hist_loadtimes_stacked}
\title{Analysis plots}
\usage{
plot_concurrency_time_by_eventtype(df)

plot_gantt(df)

plot_gantt_session(df)

plot_gantt_iteration(df)

plot_gantt_duration(df)

plot_gantt_latency(df)

plot_timestamp_time(df)

plot_timeline(df)

plot_timeline_stacked(df)

hist_loadtimes_stacked(df, max_load_time = 5)
}
\arguments{
\item{df}{dataframe returned from \code{\link{tidy_loadtest}}}
}
\description{
Analysis plots
}
