% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-log.R
\name{getErrors}
\alias{getErrors}
\title{Get errors from event log}
\usage{
getErrors(eventLog)
}
\arguments{
\item{eventLog}{Returned from \code{\link{loadTest}}. The \code{eventLog} is
a list composed of a data frame with timing from successful sessions and a
numeric with the number of failed sessions.}
}
\description{
Get errors from event log
}
\details{
Given an \code{eventLog}, this function returns the number of failed
  sessions.
}
