% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze-log.R
\name{getEventInterval}
\alias{getEventInterval}
\title{Get duration for simple event pairs}
\usage{
getEventInterval(eventLog, startEvent, endEvent, workerId = TRUE)
}
\arguments{
\item{eventLog}{Data frame of events returned from either
\code{\link{loadTest}} or \code{\link{getSuccesses}}}

\item{startEvent}{The first event in the pair, i.e. "Navigating to Shiny app"}

\item{endEvent}{The second event in the pair, i.e. "Shiny app started"}

\item{workerId}{Should the returned data frame include the worker id}
}
\value{
A data frame containing the duration of the event pair per connection
}
\description{
Get duration for simple event pairs
}
