% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{guessTimeout}
\alias{guessTimeout}
\title{Infer Timeouts}
\usage{
guessTimeout(num_concurrent, default = 15)
}
\arguments{
\item{num_concurrent}{number of concurrent connections}

\item{default}{the number of seconds to wait with no concurrency}
}
\description{
Infer Timeouts
}
\details{
Uses a hueristic to set phantomJS timeout and shiny load timeout
  based on the number of concurrent connections
}
