% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-load.R
\name{loadTest}
\alias{loadTest}
\title{Run load tests for a Shiny application}
\usage{
loadTest(testFile = "./tests/myloadtest.R", url = NULL, numConcurrent = 4,
  testDurationSec = 360)
}
\arguments{
\item{testFile}{The file containing a test script to run. Test script can be
generated from \code{recordTest(load_mode = TRUE)}.}

\item{url}{Web address of the deployed Shiny application.}

\item{numConcurrent}{Number of concurrent connections to simulate.}

\item{testDurationSec}{Total length of test. Best practice is to run tests
for > 3x the time of an individual session}
}
\value{
Returns a list with two entries, a data frame with timing information
  for successful sessions and a numeric entry with the number of failed
  sessions.
}
\description{
Run load tests for a Shiny application
}
\details{
This function simulates load against a deployed Shiny app. The
  function creates \code{numConcurrent} parallel processes that each launch a
  phantomJS process that browses to \code{url} and drives the app through the
  \code{testFile}. Tests should be generated using the \code{recordTest}
  function with \code{load_mode = TRUE}. Timing information for each process
  is written to a directory and then aggregated and returned.
  \code{numConcurrent} connections are maintained for \code{testDurationSec}
  seconds.
}
