% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-load.R
\name{loadTest}
\alias{loadTest}
\title{Run load tests for a Shiny application}
\usage{
loadTest(testFile = "./tests/myloadtest.R", url = NULL, numConcurrent = 4,
  numTotal = numConcurrent, stagger = 5)
}
\arguments{
\item{testFile}{The file containing a test script to run. Test script can be
generated from \code{recordTest(load_mode = TRUE)}.}

\item{url}{Web address of the deployed Shiny application.}

\item{numConcurrent}{Number of concurrent connections to simulate.}

\item{numTotal}{Total number of connections to simulation.}

\item{stagger}{Concurrent connections are staggered by pausing for a random,
uniformally distributed period of time before starting the connection. This
parameter is the maximum amount of time (sec) to wait. A value of 0 means
that no staggering will occur.}

\item{phantomTimeout}{Number of milliseconds to wait for the slave phantomJS
processes to start. Default to 10 seconds.}
}
\value{
If all of the child processes complete successfully this function
  returns a dataframe containing timing and event information. If not all of
  the child processes suceed, a list is returned containing timing for
  successful tests and error messages for failed tests.
}
\description{
Run load tests for a Shiny application
}
\details{
This function simulates load against a deployed Shiny app. The
  function creates a cluster of workers using the parallel function
  \code{makePSOCKcluster}. The number of works is equal to the number of
  desired concurrent connections. Each worker launches a phantomJS process
  that calls the URL and drives the app through the test. Tests should be
  generated using the \code{recordTest} function with \code{load_mode =
  TRUE}. Timing information is aggregated and returned as a data frame. If
  the number of total tests > number of concurrent tests, then finished
  workers will be recycled to complete remaining tests.The phantomJS process
  is reused, but a new browser session is started for each test.
}

