% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.R
\name{make_report}
\alias{make_report}
\title{Make shinyloadtest Report}
\usage{
make_report(df, output = "test.html", duration_cutoff = 10,
  http_latency_cutoff = 5, max_websocket_cutoff = 20,
  open_browser = TRUE, self_contained = TRUE, verbose = TRUE)
}
\arguments{
\item{df}{data.frame returned from \code{\link{tidy_loadtest}}}

\item{output}{File where HTML output should be saved}

\item{duration_cutoff}{Cutoff value for session duration plot}

\item{http_latency_cutoff}{Cutoff value for total http latency plot}

\item{max_websocket_cutoff}{Cutoff value for max websocket latency plot}

\item{open_browser}{Whether to open the created output in the browser}

\item{self_contained}{Boolean that determines if the final output should be a self contained html file}

\item{verbose}{Boolean that determines if progress output is displayed}
}
\description{
Make shinyloadtest Report
}
