% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{tidy_loadtest}
\alias{tidy_loadtest}
\title{Create Tidy Load Test Results}
\usage{
tidy_loadtest(...)
}
\arguments{
\item{...}{Key-value pairs where the key is the desired name for the test and the
value is a path to the test result directory.}
}
\value{
A tidy data frame with the test result data. Each row is an event. Columns include
   identifiers and timing information for the event.
}
\description{
The \code{shinycannon} tool creates a directory of log files for
  each load test. This function translates one or more test result
  directories into a tidy data frame.
}
\examples{
\dontrun{
  tidy_loadtest(
     `1 core` = 'results/test-1/',
     `2 cores` = 'results/test-2/'
  )
}
}
